Copy java9rtexport changes from sbt/sbt-launcher-package and update repository/organization/artifact information to use org.scala-sbt
