+++ README.md
@@ -1,10 +1,21 @@
 ## Java 9 RT Exporter
 
-Usage:
+### Compile
 
 ```
-% java_use 1.8
+$ java_use 1.8
 $ sbt clean package
+```
+
+### Print recommended directory
+
+```
+$ java -jar target/java9-rt-export-*.jar --rt-ext-dir
+```
+
+### Export rt.jar
+
+```
 $ java_use 9
-$ java -jar target/java9-rt-export-*.jar $HOME/.sbt/0.13/rt.jar)
+$ java -jar target/java9-rt-export-*.jar $HOME/.sbt/java9-rt-ext/rt.jar
 ```
